+++ examples/allocations.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use scylla::transport::session::Session;
+use scylla::session::Session;
 use scylla::{statement::prepared_statement::PreparedStatement, SessionBuilder};
 use std::io::Write;
 use std::sync::atomic::{AtomicUsize, Ordering};
