+++ examples/cqlsh-rs.rs
@@ -4,8 +4,8 @@ use rustyline::error::ReadlineError;
 use rustyline::{CompletionType, Config, Context, Editor};
 use rustyline_derive::{Helper, Highlighter, Hinter, Validator};
 use scylla::frame::response::result::Row;
+use scylla::session::Session;
 use scylla::transport::query_result::IntoRowsResultError;
-use scylla::transport::session::Session;
 use scylla::transport::Compression;
 use scylla::QueryResult;
 use scylla::SessionBuilder;
