+++ examples/execution_profile.rs
@@ -2,9 +2,9 @@ use anyhow::Result;
 use scylla::load_balancing;
 use scylla::query::Query;
 use scylla::retry_policy::{DefaultRetryPolicy, FallthroughRetryPolicy};
+use scylla::session::Session;
 use scylla::speculative_execution::PercentileSpeculativeExecutionPolicy;
 use scylla::statement::{Consistency, SerialConsistency};
-use scylla::transport::session::Session;
 use scylla::transport::ExecutionProfile;
 use scylla::{SessionBuilder, SessionConfig};
 use std::env;
