+++ scylla/src/session/mod.rs
@@ -1,0 +1,10 @@
+#[allow(clippy::module_inception)]
+mod session;
+pub mod session_builder;
+#[cfg(test)]
+mod session_test;
+
+pub use session::*;
+#[cfg(feature = "cloud")]
+pub use session_builder::CloudSessionBuilder;
+pub use session_builder::SessionBuilder;
