+++ scylla/src/transport/mod.rs
@@ -1,7 +1,7 @@
 pub(crate) mod caching_session;
-mod cluster;
+pub(crate) mod cluster;
 pub(crate) mod connection;
-mod connection_pool;
+pub(crate) mod connection_pool;
 pub mod downgrading_consistency_retry_policy;
 pub mod errors;
 pub mod execution_profile;
@@ -12,21 +12,16 @@ pub mod load_balancing;
 pub mod locator;
 pub mod metadata;
 pub(crate) mod metrics;
-mod node;
+pub(crate) mod node;
 pub mod partitioner;
 pub mod query_result;
 pub mod retry_policy;
-pub mod session;
-pub mod session_builder;
 pub mod speculative_execution;
 
 pub use crate::frame::{Authenticator, Compression};
 pub use connection::SelfIdentity;
 pub use execution_profile::ExecutionProfile;
 pub use scylla_cql::frame::request::query::{PagingState, PagingStateResponse};
-
-#[cfg(test)]
-mod session_test;
 
 pub use cluster::ClusterData;
 pub use node::{KnownNode, Node, NodeAddr, NodeRef};
