+++ scylla/tests/integration/utils.rs
@@ -3,8 +3,8 @@ use scylla::deserialize::DeserializeValue;
 use scylla::load_balancing::{FallbackPlan, LoadBalancingPolicy, RoutingInfo};
 use scylla::query::Query;
 use scylla::routing::Shard;
+use scylla::session::session_builder::{GenericSessionBuilder, SessionBuilderKind};
 use scylla::transport::errors::QueryError;
-use scylla::transport::session_builder::{GenericSessionBuilder, SessionBuilderKind};
 use scylla::transport::{ClusterData, NodeRef};
 use scylla::{ExecutionProfile, Session};
 use std::collections::HashMap;
@@ -143,7 +143,7 @@ pub(crate) fn create_new_session_builder() -> GenericSessionBuilder<impl Session
 
         #[cfg(scylla_cloud_tests)]
         {
-            use scylla::transport::session_builder::CloudMode;
+            use scylla::session::session_builder::CloudMode;
             use scylla::CloudSessionBuilder;
             use std::path::Path;
